LZ-21 Moved download buttons to surveys list
